@@ -9,7 +9,9 @@
     "build": "babel . --out-dir build",
     "start:prod": "NODE_ENV=production nodemon server.js",
     "debug": "ndb server.js",
-    "heroku-postbuild": "npm install --production"
+    "heroku-postbuild": "npm install --production",
+    "test": "mocha ./tests/ --require @babel/register -r dotenv/config --timeout 30000 --exit ",
+    "coverage": "nyc npm run test "
   },
   "author": "Koinstreet",
   "license": "ISC",
@@ -21,26 +23,43 @@
     "@babel/polyfill": "^7.12.1",
     "@babel/preset-env": "^7.12.11",
     "@babel/runtime": "^7.12.5",
+    "bcrypt": "^5.0.1",
     "bcryptjs": "^2.4.3",
+    "chai": "^4.3.4",
+    "chai-http": "^4.3.0",
     "cloudinary": "^1.21.0",
     "cookie-session": "^1.4.0",
     "cors": "^2.8.5",
+    "crypto": "^1.0.1",
     "dotenv": "^8.2.0",
+    "eth-sig-util": "^3.0.1",
+    "ethereumjs-util": "^7.1.4",
+    "ethers": "^5.5.3",
     "express": "^4.17.1",
+    "google-auth-library": "^7.5.0",
     "http-status-codes": "^1.4.0",
+    "ipfs-http-client": "^50.1.2",
     "jsonwebtoken": "^8.5.1",
     "loadash": "^1.0.0",
     "lodash": "^4.17.15",
-    "mongoose": "^5.11.18",
+    "mocha": "^9.0.3",
+    "moment": "^2.29.1",
+    "mongodb": "^4.0.1",
+    "mongoose": "^5.13.9",
     "morgan": "^1.10.0",
     "multer": "^1.4.2",
     "nodemailer": "^6.6.2",
+    "nyc": "^15.1.0",
     "passport": "^0.4.1",
     "passport-facebook": "^3.0.0",
+    "passport-facebook-token": "^4.0.0",
     "passport-github": "^1.1.0",
     "passport-google-oauth20": "^2.0.0",
     "passport-linkedin": "^1.0.0",
     "passport-linkedin-oauth2": "^2.0.0",
+    "python-shell": "^3.0.1",
+    "request": "^2.88.2",
+    "socket.io": "^4.1.3",
     "validator": "^13.0.0"
   },
   "devDependencies": {
@@ -58,6 +77,7 @@
     "prettier": "^1.19.1"
   },
   "engines": {
-    "node": ">=10.0.0"
+    "node": "16.x",
+    "npm": "6.x"
   }
 }
