@@ -1,9 +1,15 @@
 const { CREATED, UNAUTHORIZED, BAD_REQUEST, OK } = require("http-status-codes");
+const moment = require("moment");
+const url = require("url");
+const querystring = require("querystring");
 
 // DB
 const Event = require("../../../model/v1/Events");
 
-const validateEvent  = require("../../../validators/event");
+const validateEvent = require("../../../validators/event");
+
+import sendEmail from "../../../utils/email/sendEmail";
+import emailTemplate from "../../../utils/email/event/createEvent";
 
 const uploadImage = require("../../../utils/uploadImage");
 
@@ -37,7 +43,21 @@ exports.createEvent = async (req, res, next) => {
       };
     }
 
-    const newEvent= await Event.create(event);
+    const newEvent = await Event.create(event);
+
+    const subject = "Event Successfuly Created!";
+    sendEmail(
+      emailTemplate(
+        req.user.firstName,
+        newEvent.EventPicture,
+        newEvent.eventName,
+        newEvent.EventDescription,
+        newEvent.eventLink
+      ),
+      subject,
+      req.user.email
+    );
+
     return successWithData(
       res,
       CREATED,
@@ -52,7 +72,46 @@ exports.createEvent = async (req, res, next) => {
 
 exports.getAllEvents = async (req, res, next) => {
   try {
-    const events = await Event.find({})
+    let parsedUrl = url.parse(req.url);
+    let parsedQs = querystring.parse(parsedUrl.query);
+
+    let time = 100;
+
+    if (parsedQs.event_time === "week") {
+      time = 7;
+    } else if (parsedQs.event_time === "month") {
+      time = 30;
+    } else {
+      time = 100;
+    }
+    let timeFilter = parsedQs.event_time
+      ? {
+          time: {
+            $gte: new Date(
+              moment()
+                .subtract(time, "days")
+                .startOf("day")
+                .format("YYYY-MM-DD HH:mm:ss")
+            ),
+            $lt: new Date(),
+          },
+        }
+      : {};
+    let virtualFilter = parsedQs.Virtual ? { Virtual: parsedQs.Virtual } : {};
+    let featuredFilter = parsedQs.Featured
+      ? { Featured: parsedQs.Featured }
+      : {};
+    let catNameFilter = parsedQs.catName
+      ? { catName: { $in: parsedQs.catName.split(",") } }
+      : {};
+
+    let allFilter = {
+      $and: [timeFilter, virtualFilter, featuredFilter, catNameFilter],
+    }
+      ? { $and: [timeFilter, virtualFilter, featuredFilter, catNameFilter] }
+      : {};
+
+    const events = await Event.find(allFilter)
       .populate("host", "-password")
       .sort("-createdAt");
     return successWithData(res, OK, "Events fetched successfully", events);
@@ -68,7 +127,10 @@ exports.getEvent = async (req, res, next) => {
       "host",
       "-password"
     );
-    if (!event) return AppError.tryCatchError(res, err);
+    if (!event) {
+      let error = { message: "undefined event" };
+      return AppError.tryCatchError(res, error);
+    }
     return successWithData(res, OK, "Event fetched successfully", event);
   } catch (err) {
     console.log(err);
@@ -77,7 +139,6 @@ exports.getEvent = async (req, res, next) => {
 };
 
 exports.updateEvent = async (req, res, next) => {
-  
   try {
     const { errors, isValid } = validateEvent(req.body);
     if (!isValid) {
@@ -85,7 +146,11 @@ exports.updateEvent = async (req, res, next) => {
     }
 
     const eventUpdate = await Event.findById(req.params.id);
-    if (!eventUpdate) return AppError.tryCatchError(res, err);
+    if (!eventUpdate) {
+      let error = { message: "undefined event" };
+      return AppError.tryCatchError(res, error);
+    }
+
     let event;
     if (req.file) {
       const data = await uploadImage(req.file);
