@@ -64,12 +64,9 @@ exports.getAllLocations = async (req, res, next) => {
 
 exports.getLocation = async (req, res, next) => {
   try {
-    const Location = await Location.findById(req.params.id).populate(
-      "added_by",
-      "-password"
-    );
-    if (!Location) return AppError.tryCatchError(res, err);
-    return successWithData(res, OK, "Location fetched successfully", Location);
+    const location = await Location.findById(req.params.id).populate("added_by").sort("-createdAt");
+    if (!location) { let error = {message: "undefined location"}; return AppError.tryCatchError(res, error);}
+    return successWithData(res, OK, "locationfetched successfully", location);
   } catch (err) {
     console.log(err);
     return AppError.tryCatchError(res, err);
@@ -85,7 +82,8 @@ exports.updateLocation = async (req, res, next) => {
     }
 
     const LocationUpdate = await Location.findById(req.params.id);
-    if (!LocationUpdate) return AppError.tryCatchError(res, err);
+    if (!LocationUpdate) { let error = {message: "undefined Location"}; return AppError.tryCatchError(res, error);}
+
     let location;
     if (req.file) {
       const logo = await uploadImage(req.file);
@@ -104,7 +102,7 @@ exports.updateLocation = async (req, res, next) => {
     }
     const modifiedLocation = await Location.findOneAndUpdate(
       { _id: req.params.id },
-      { ...Location },
+      { ...location },
       { new: true }
     );
     return successWithData(res, OK, "Location modified", modifiedLocation);
