@@ -4,6 +4,8 @@ const { CREATED, UNAUTHORIZED, BAD_REQUEST, OK } = require("http-status-codes");
 const Services = require("../../../model/v1/service");
 
 const validateService = require("../../../validators/service");
+import sendEmail from '../../../utils/email/sendEmail';
+import emailTemplate from '../../../utils/email/consultancy/createProject';
 
 const {
   successWithData,
@@ -13,6 +15,8 @@ const {
 // Error
 const AppError = require("../../../utils/appError");
 
+const Notifications = require("../../../model/v1/notifications");
+
 exports.createService = async (req, res, next) => {
   try {
     const { errors, isValid } = validateService(req.body);
@@ -21,9 +25,16 @@ exports.createService = async (req, res, next) => {
     }
     let service = {
         ...req.body,
-      };
+        authorId: req.user.id,
+      };    
 
     const newService= await Services.create(service);
+
+    const subject = `Your project has been submitted, you will be notified once your project is claimed`;
+
+    sendEmail(emailTemplate(req.user.firstName, req.body.project_name ? req.body.project_name : '', req.body.project_details ? req.body.project_details : '', req.body.launch_date ? req.body.launch_date : ''), subject, req.user.email);
+    const newNotification = await Notifications.create({receiverId: req.user.id, title: subject, project_id: newService._id, type: 'Project', authorId: null});
+
     return successWithData(
       res,
       CREATED,
@@ -76,6 +87,7 @@ exports.updateService = async (req, res, next) => {
 
     let service = {
         ...req.body,
+        authorId: req.user.id,
       };
     
     const modifiedService = await Services.findOneAndUpdate(
