@@ -13,8 +13,9 @@ const LocationSchema = new Schema({
   },
   LocationName: {
     type: String,
+    required: true,
   },
-  title: {
+  location: {
     type: String,
     required: true,
   },
@@ -26,8 +27,12 @@ const LocationSchema = new Schema({
   },
   description: {
     type: String,
+    required: true
   },
   chapter_leader: {
+    type: String,
+  },
+  chapter_type: {
     type: String,
   },
   member_size: {
