@@ -13,17 +13,27 @@ const moduleSchema = new Schema({
     required: true,
     ref: "Course",
   },
+  image: {
+    type: String
+  },
+  url: {
+      type: String
+  },
   name: {
     type: String,
-    required: true,
   },
-  type: {
+  
+  level: {
     type: String,
-    required: true,
+  },
+  duration: {
+    type: String
+  },
+  description: {
+    type: String
   },
   content: {
     type: Object,
-    required: true,
   },
 },
 {
