@@ -3,11 +3,10 @@ const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
 
 const serviceSchema = new Schema({
-  cost: {
-    type: Number,
-  },
-  amount: {
-    type: Number,
+  authorId: {
+    type: mongoose.Schema.Types.ObjectId,
+    required: true,
+    ref: "User",
   },
   project_name: {
     type: String,
@@ -16,7 +15,38 @@ const serviceSchema = new Schema({
   industry: {
     type: String,
   },
-  email: {
+  project_budget: {
+    type: String,
+  },
+  budget_currency: {
+    type: String,
+  },
+  project_stacks: {
+    type: Array,
+  },
+  PM_fullName: {
+    type: String,
+  },
+  Agree_terms: {
+    type: Boolean,
+    default: false,
+  },
+  project_size: {
+    type: String,
+  },
+  project_details: {
+    type: String,
+  },
+  webiste: {
+    type: String,
+  },
+  project_tags: {
+    type: Array,
+  },
+  PM_email: {
+    type: String,
+  },
+  contact_email: {
     type: String,
     required: true,
   },
@@ -24,7 +54,14 @@ const serviceSchema = new Schema({
     type: Number,
   },
   launch_date: {
-    type: Date,
+    type: String,
+  },
+  pay_option: {
+    type: String,
+  },
+  claimed: {
+    type: Boolean,
+    default: false
   },
 }, {
   timestamps: true
