@@ -9,20 +9,27 @@ const authMiddleware = require("../middleware/auth");
 
 const router = express.Router();
 
-router.get("/", jobs.getAllJobs);
 
-router.get("/:id", jobs.getJob);
+router.get("/userJobs",authMiddleware.protect, jobs.getUserJobs);
+router.get("/candidates",authMiddleware.protect, jobs.getCandidates);
 
-router.use(authMiddleware.protect);
+
 
 router.post(
-  "/",
+  "/",authMiddleware.protect,
   jobs.createJob
 );
 router
   .route("/:id")
-  .put(jobs.updateJob)
-  .delete(jobs.deleteJob);
+  .patch(authMiddleware.protect, jobs.updateJob)
+  .delete(authMiddleware.protect, jobs.deleteJob);
+
+
+router.post("/search", jobs.getSearch);
+
+router.get("/", jobs.getAllJobs);
+
+router.get("/:id", jobs.getJob);
 
 
 module.exports = router;
