@@ -1,11 +1,11 @@
 const express = require("express");
 
-
 // NEW CONTROLLERS
 const mentee = require("../controllers/v1/mentee");
 
 // MIDDLEWARE
 const authMiddleware = require("../middleware/auth");
+const fileUploader = require("../middleware/fileUploaders");
 
 const router = express.Router();
 
@@ -15,14 +15,10 @@ router.get("/:id", mentee.getMentee);
 
 router.use(authMiddleware.protect);
 
-router.post(
-  "/",
-  mentee.createMentee
-);
+router.post("/", fileUploader.uploadQRImage, mentee.createMentee);
 router
   .route("/:id")
-  .put(mentee.updateMentee)
+  .patch(fileUploader.uploadQRImage, mentee.updateMentee)
   .delete(mentee.deleteMentee);
-
 
 module.exports = router;
