@@ -1,11 +1,11 @@
 const express = require("express");
 
-
 // NEW CONTROLLERS
 const mentor = require("../controllers/v1/mentor");
 
 // MIDDLEWARE
 const authMiddleware = require("../middleware/auth");
+const fileUploader = require("../middleware/fileUploaders");
 
 const router = express.Router();
 
@@ -15,14 +15,10 @@ router.get("/:id", mentor.getMentor);
 
 router.use(authMiddleware.protect);
 
-router.post(
-  "/",
-  mentor.createMentor
-);
+router.post("/", fileUploader.uploadQRImage, mentor.createMentor);
 router
   .route("/:id")
-  .put(mentor.updateMentor)
+  .patch(fileUploader.uploadQRImage, mentor.updateMentor)
   .delete(mentor.deleteMentor);
-
 
 module.exports = router;
