@@ -6,7 +6,7 @@ const subscriberController = require("../controllers/v1/subscriber");
 const router = express.Router();
 
 router
-.get('/', subscriberController.getAllSubscriber)
-.post('/', subscriberController.addSubscriber);
+  .get("/", subscriberController.getAllSubscriber)
+  .post("/", subscriberController.addSubscriber);
 
 module.exports = router;
