@@ -3,28 +3,16 @@ const _ = require("lodash");
 
 const validateEvent = (data) => {
   let errors = {};
-  data.catName = data.catName ? data.catName : "";
+
   data.eventName = data.eventName ? data.eventName : "";
   data.eventLink = data.eventLink ? data.eventLink : "";
-  data.actionLink = data.actionLink ? data.actionLink : "";
-  data.callToAction = data.callToAction ? data.callToAction : "";
   data.EventDescription = data.EventDescription ? data.EventDescription : "";
 
-
-  if (validator.isEmpty(data.catName)) {
-    errors.catName = "catName is required";
-  }
   if (validator.isEmpty(data.eventName)) {
     errors.eventName = "eventName field is required";
   }
   if (validator.isEmpty(data.eventLink)) {
     errors.eventLink = "eventLink field is required";
-  }
-  if (validator.isEmpty(data.actionLink)) {
-    errors.actionLink = "actionLink is required";
-  }
-  if (validator.isEmpty(data.callToAction)) {
-    errors.callToAction = "callToAction is required";
   }
   if (validator.isEmpty(data.EventDescription)) {
     errors.EventDescription = "EventDescription is required";
